two ways to destroy your cluster
the cluster can be destroyed, or the parent resource group can be
destroyed with the cluster in it

(Both are big disasters, although we can protect against through
rigorous policy and tenant isolation, there is a possibility that
remains, so we have to be prepared for both)

This is the subject of my KubeCon talk! What do you do when your RG has
been deleted, how can you be prepared for that?

Signed-off-by: Kingdon Barrett <kingdon@weave.works>